Refactor ARCameraPageAndroid
